#265: Added support for stops in size legend
