+++ examples/index.html
@@ -26,6 +26,6 @@
         <div id="chart3"></div>
 </div>
     <!-- <script /src="https://d3js.org/d3.v4.js"></script> -->
-    <script src="./js/wildfires.js"></script>
+    <script src="./js/sample3.js"></script>
 </body>
 </html>
