+++ packages/muze-axis/src/size-axis/defaults.js
@@ -1,4 +1,5 @@
 export const DEFAULT_CONFIG = {
     value: 36,
-    range: [36, 1000]
+    range: [36, 1000],
+    stops: 2
 };
